ISIS-1976: let PersistenceSession no longer implement AdapterManager
Task-Url: https://issues.apache.org/jira/browse/ISIS-1976
