+++ core/viewer-wicket-impl/src/main/java/org/apache/isis/viewer/wicket/viewer/integration/wicket/ConverterForObjectAdapterMemento.java
@@ -25,6 +25,7 @@ import com.google.common.base.Strings;
 
 import org.apache.wicket.util.convert.IConverter;
 
+import org.apache.isis.commons.internal.exceptions._Exceptions;
 import org.apache.isis.core.metamodel.adapter.ObjectAdapter;
 import org.apache.isis.core.metamodel.adapter.mgr.AdapterManager.ConcurrencyChecking;
 import org.apache.isis.core.metamodel.adapter.oid.Oid;
@@ -54,8 +55,11 @@ public class ConverterForObjectAdapterMemento implements IConverter<ObjectAdapte
             return null;
         }
         final Oid oid = RootOid.deStringEncoded(value);
-        final ObjectAdapter adapter = getPersistenceSession().lookupAdapterFor(oid);
-        return ObjectAdapterMemento.createOrNull(adapter);
+        
+        //FIXME[ISIS-1976]
+        //final ObjectAdapter adapter = getPersistenceSession().lookupAdapterFor(oid);
+        //return ObjectAdapterMemento.createOrNull(adapter);
+        throw _Exceptions.notImplemented();
     }
 
     /**
